change main package
change main package to use repo name
