+++ internal/models/user.go
@@ -1,8 +1,11 @@
 package models
 
+import "gorm.io/gorm"
+
 type User struct {
-    ID          int `gorm:"primaryKey"`
-    Username    string
-    Email       string
-    Password    string
+	gorm.Model
+
+	Username string `gorm:"unique" json:"username"`
+	Email    string `gorm:"unique" json:"email"`
+	Password string `gorm:"not null" json:"password"`
 }
